back/requirements.txt
click==7.1.2
cycler==0.10.0
Cython==0.29.20
Flask==1.1.2
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
torch==1.5.1
torchvision==0.6.1
tqdm==4.47.0
Werkzeug==1.0.1
